refactor: dependency-rule rewrite to typescript, using `@typescript-eslint/utils`, defining helper types
